优化packer， 不传入msgID

+++ pkg/network/message.go
@@ -3,7 +3,7 @@ package network
 type Message struct {
 	ID    uint64
 	Data  []byte
-	RawID uint64
+	RawID uint32
 }
 
 type Packet struct {
